Russify DPS abbreviation entries
