+++ assets/templates/abbreviation-sbs.html
@@ -1,35 +1,23 @@
 <%page args='abbreviation'/>
 <%namespace file='/lib.mako' import='google_link_sbs'/>
-<%
-    from datetime import date
-
-    today = date.today()
-
-    abbrev = abbreviation.iloc[0]
-    meaning = abbreviation.iloc[1]
-    pali_meaning = abbreviation.iloc[2]
-    ru_meaning = abbreviation.iloc[3]
-    examp = abbreviation.iloc[4]
-    expl = abbreviation.iloc[5]
-%>
 <body>
     <div class="help">
         <p>
-            abbreviation. <b>${abbrev}</b>. ${meaning}.\
-%if pali_meaning != '':
- ${pali_meaning}.\
+            abbreviation. <b>${abbreviation}</b>. ${abbreviation.meaning}.\
+%if abbreviation.pali_meaning != '':
+ ${abbreviation.pali_meaning}.\
 %endif
             <br>
-            %if examp != '':
-            e.g. ${examp}.
+            %if abbreviation.example != '':
+            e.g. ${abbreviation.example}.
             %endif
             <br>
-            %if expl != '':
-            ${expl}.
+            %if abbreviation.explanation != '':
+            ${abbreviation.explanation}.
             %endif
 
     <p>
-        ${google_link_sbs(entry=abbrev, text='Report a mistake.')}
+        ${google_link_sbs(entry=abbreviation, text='Report a mistake.')}
     </p>
 
 ## vi: syn=mako
