+++ html_components.py
@@ -11,6 +11,8 @@ from pandas.core.frame import Series
 
 import helpers
 
+from word import DpsWord
+
 
 class TemplateBase:
     def __init__(self, template_path: Path):
@@ -36,7 +38,7 @@ class HeaderTemplate(TemplateBase):
 
 
 class WordTemplate(TemplateBase):
-    def render(self, word: helpers.DpsWord, table_data_read: str) -> str:
+    def render(self, word: DpsWord, table_data_read: str) -> str:
         return self._render_helper(
             conjugations=helpers.CONJUGATIONS,
             declensions=helpers.DECLENSIONS,
